[mypyc] Refactor: extract non-local control classes from mypyc.genops
